Added functionality to fetch Metro API data, corrected type in MetroResponse struct, corrected MetroAPI url

+++ utils/structs.go
@@ -36,7 +36,7 @@ type RestCountriesResponse struct {
 */
 type MetroResponse struct {
 	Hourly struct {
-		Precipitation []int       	`json:"precipitation"`
+		Precipitation []float64     `json:"precipitation"`
 		Temperature2M []float64    	`json:"temperature_2m"`
 	} 								`json:"hourly"`
 }
